refactor(blog): extract content filtering logic to utility function
