+++ apps/blog/src/stores/contents-store.ts
@@ -2,38 +2,22 @@
 
 import { atom, createStore } from 'jotai'
 
-import { type Memo, type Post, memo, post } from '#site/content'
+import type { Memo, Post } from '#site/content'
+import { getContents } from '#utils/contents'
 
 export type Filter = 'all' | 'post' | 'memo' | undefined
 export type SelectedFilter = 'all' | 'post' | 'memo'
 export type Content = Memo & Post
 
-const toSortedByDate = (contents: (Memo & Post)[]) => {
-  const contents_copy = contents.slice()
-  contents_copy.sort((a, b) => Date.parse(b.date) - Date.parse(a.date))
-  return contents_copy
-}
-
 const contentsStore = createStore()
 const filterAtom = atom<Filter>()
 const contentsAtom = atom<Content[]>()
 contentsStore.set(filterAtom, undefined)
-contentsStore.set(contentsAtom, toSortedByDate([...memo, ...post]))
+contentsStore.set(contentsAtom, getContents())
 
 const unsub = contentsStore.sub(filterAtom, () => {
   const filter = contentsStore.get(filterAtom)
-
-  switch (filter) {
-    case 'post':
-      contentsStore.set(contentsAtom, toSortedByDate(post))
-      break
-    case 'memo':
-      contentsStore.set(contentsAtom, toSortedByDate(memo))
-      break
-    default:
-      contentsStore.set(contentsAtom, toSortedByDate([...memo, ...post]))
-      break
-  }
+  contentsStore.set(contentsAtom, getContents(filter))
 })
 
 export { contentsStore, contentsAtom, filterAtom, unsub }
